requirements.txt
streamlit==1.45.0
aiohttp==3.9.5
requests==2.32.3
pillow==11.0.0
beautifulsoup4==4.12.3
lxml==5.3.0
pymupdf==1.24.14
playwright==1.57.0
tqdm==4.67.1
